% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_mnb_color.R
\name{closest_mnb_color}
\alias{closest_mnb_color}
\title{Find the closest MNB Color}
\usage{
closest_mnb_color(colors, limit = 300)
}
\arguments{
\item{colors}{A character vector of color names or hexadecimal codes.}

\item{limit}{Numeric. The maximum RGB distance allowed for a match. Default is 200.}
}
\value{
A character vector of the closest matching MNB colors or the original color if no match is found within the limit.
}
\description{
This function takes a vector of colors and returns the closest matching color from the MNB color palette.
}
\examples{
closest_mnb_color("red")
closest_mnb_color(c("blue", "green", "#FF00FF"))

}
