% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{.bigmark}
\alias{.bigmark}
\title{Determine Big Mark for Number Formatting}
\usage{
.bigmark()
}
\value{
A logical value: TRUE if the language is Hungarian or not set, FALSE otherwise.
}
\description{
This function determines the appropriate big mark for number formatting based on the language setting.
}
\keyword{internal}
