% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.R
\name{geom_text_box}
\alias{geom_text_box}
\title{Add Formatted Text Box to ggplot}
\usage{
geom_text_box(
  label = "Text \\n[blue] blue",
  ...,
  size = 5,
  base_color = "black",
  base_fill = "transparent",
  label_size = 1,
  text.color = "black",
  family = "Cambria",
  closest_color = TRUE
)
}
\arguments{
\item{label}{Text to display in the box. Can include color formatting using square brackets, e.g., "\link{blue}Blue text[]".}

\item{...}{Additional arguments passed to ggtext::geom_richtext().}

\item{size}{Text size. Default is 5.}

\item{base_color}{Color of the box border. Default is "black".}

\item{base_fill}{Fill color of the box. Default is "transparent".}

\item{label_size}{Size of the label. Default is 1.}

\item{text.color}{Color of the text. Default is "black".}

\item{family}{Font family. Default is "Cambria".}

\item{closest_color}{Logical. If TRUE, uses the closest MNB color for color names. Default is TRUE.}
}
\value{
A ggplot2 layer object.
}
\description{
This function adds a formatted text box to a ggplot using the ggtext package.
}
\examples{
library(ggplot2)
ggplot() +
  geom_text_box(label = "Hello [blue]World[]!", x = 0, y = 0)

}
