% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfd_data.R
\name{gfd_data}
\alias{gfd_data}
\title{Downloads data from GFD API.}
\usage{
gfd_data(series, ..., skip_rows = 2)
}
\arguments{
\item{series}{A string with the series ID.}

\item{...}{Additional arguments to be passed to the API. (i.e.: periodicity = "Monthly")}
}
\value{
A data.frame with the downloaded data.
}
\description{
The function uses the GFD API. You need to set the environment variable \code{GFD_USER} and \code{GFD_PWD} with your username and password.
}
\examples{
gfd_data("_SPXTRD")

}
