% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt_mnb}
\alias{gt_mnb}
\title{Create an MNB-styled GT Table}
\usage{
gt_mnb(data, title = NULL, comment = NULL)
}
\arguments{
\item{data}{A data frame or GT table object to be styled.}

\item{title}{Optional. A string to be used as the table title.}

\item{comment}{Optional. A string to be added as a source note to the table.}
}
\value{
A GT table object with MNB styling applied.
}
\description{
This function creates a GT table with MNB (Magyar Nemzeti Bank) styling applied.
It formats numbers, percentages, and applies specific styles to the table.
}
\examples{
iris |> 
  head() |> 
  gt_mnb(title = "Random title", comment = "Source: This is a built-in dataset")

}
