% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnb_colors.R
\name{mnb_colors}
\alias{mnb_colors}
\title{MNB Color Palette}
\usage{
mnb_colors(n = NULL)
}
\arguments{
\item{n}{Optional. Number of colors to return. If NULL (default), all colors are returned.}
}
\value{
A character vector of color hexadecimal codes.
}
\description{
This function returns a vector of colors used in the MNB (Magyar Nemzeti Bank) color palette.
}
\examples{
# Get all MNB colors
mnb_colors()

}
