% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{mnb_divergent_colors}
\alias{mnb_divergent_colors}
\title{MNB Divergent Colors}
\usage{
mnb_divergent_colors(n, rev = FALSE)
}
\arguments{
\item{n}{Number of colors}

\item{rev}{Logical. If TRUE, reverses the order of the colors. Default is FALSE.}
}
\value{
A vector of colors
}
\description{
This function returns the divergent colors for a given n.
}
\examples{
library(ggplot2)
scales::show_col(mnb_divergent_colors(4))

}
