% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_font.R
\name{modify_svg_fonts}
\alias{modify_svg_fonts}
\title{Modify SVG Fonts}
\usage{
modify_svg_fonts(
  input_file,
  output_file = NULL,
  family = "Calibri",
  font_color = "#333333"
)
}
\arguments{
\item{input_file}{A character string specifying the path to the input SVG file.}

\item{output_file}{A character string specifying the path to the output SVG file. If NULL, the input file will be overwritten.}

\item{family}{A character string specifying the font family to use. Default is "Calibri".}

\item{font_color}{A character string specifying the font color to use. Default is "#333333".}
}
\value{
None. The function writes the modified SVG content to the output file.
}
\description{
This function modifies the font family and font color in an SVG file.
}
