% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_mnb.R
\name{scale_color_discrete_mnb_divergent}
\alias{scale_color_discrete_mnb_divergent}
\title{MNB divergent colors for ggplot2}
\usage{
scale_color_discrete_mnb_divergent()
}
\value{
A ggplot2 color scale object.
}
\description{
This function creates a color scale using the MNB color palette for ggplot2.
}
\examples{
library(ggplot2)
ggplot2::ggplot(iris) + 
  ggplot2::aes(Sepal.Length, Sepal.Width, color = Species) +
  ggplot2::geom_point(size = 3) + 
  scale_color_discrete_mnb_divergent()

}
