% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_mnb_divergent_b}
\alias{scale_color_mnb_divergent_b}
\title{MNB Divergent Binned Color Scale for ggplot2}
\usage{
scale_color_mnb_divergent_b(
  n.breaks = NULL,
  breaks = waiver(),
  rev = FALSE,
  ...
)
}
\arguments{
\item{n.breaks}{Optional. Number of breaks for the binned scale. If NULL, the number of breaks is determined by the \code{breaks} parameter.}

\item{breaks}{Optional. A vector of break points for the binned scale. Default is waiver(), which lets ggplot2 determine the breaks.}

\item{rev}{Logical. If TRUE, reverses the order of the colors. Default is FALSE.}

\item{...}{Additional arguments passed to ggplot2::binned_scale().}
}
\value{
A ggplot2 binned fill scale object using the MNB divergent color palette.
}
\description{
This function creates a binned fill scale using the MNB (Magyar Nemzeti Bank) divergent color palette for ggplot2.
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, fill = Petal.Length)) + 
  geom_point() +
  scale_color_mnb_divergent_b(n.breaks = 5, rev = TRUE)

}
