% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_number}
\alias{scale_x_number}
\title{Scale X-axis with Formatted Numbers}
\usage{
scale_x_number(..., expand = c(0, 0))
}
\value{
A ggplot2 scale object for the x-axis with formatted numbers.
}
\description{
This function creates a continuous x-axis scale with numbers formatted according to the current language setting.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  scale_x_number()

}
