% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_quarter}
\alias{scale_x_quarter}
\title{Scale X-axis with Quarter Labels}
\usage{
scale_x_quarter(..., expand = c(0, 0))
}
\value{
A ggplot2 scale object for the x-axis with quarter labels.
}
\description{
This function creates a continuous x-axis scale with labels at each quarter.
}
