% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_y_percent}
\alias{scale_y_percent}
\title{Scale Y-axis with Formatted Percent}
\usage{
scale_y_percent(..., expand = c(0, 0))
}
\value{
A ggplot2 scale object for the y-axis with formatted numbers.
}
\description{
This function creates a continuous y-axis scale with numbers formatted according to the current language setting.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  scale_y_number()

}
