% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.R
\name{text_box}
\alias{text_box}
\title{Add Formatted Text Box to ggplot}
\usage{
text_box(label, size = 5, ...)
}
\arguments{
\item{label}{Text to display in the box. Can include color formatting using square brackets, e.g., "\link{blue}Blue text[]".}

\item{size}{Text size. Default is 5.}

\item{...}{Additional arguments passed to ggtext::geom_richtext().}
}
\value{
A ggplot2 layer object.
}
\description{
This function adds a formatted text box to a ggplot using the ggtext package.
}
\examples{
library(ggplot2)
ggplot() +
  text_box(label = "Hello [blue]World[]!", x = 0, y = 0)

}
