% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box.R
\name{text_boxes}
\alias{text_boxes}
\title{Create Multiple Text Boxes}
\usage{
text_boxes(
  chars = 30,
  lines = 3,
  size = 15,
  base_color = "black",
  base_fill = "transparent",
  labels = c("[10]bluedmoewndo3[][red]\\nred []", "[40]bluedewm[][red]\\nred []")
)
}
\arguments{
\item{chars}{Number of characters for the base text. Default is 30.}

\item{lines}{Number of lines for the base text. Default is 3.}

\item{size}{Text size. Default is 15.}

\item{base_color}{Color of the box border. Default is "black".}

\item{base_fill}{Fill color of the box. Default is "transparent".}

\item{labels}{Vector of labels for the text boxes. Default includes two example labels.}
}
\value{
A ggplot object containing multiple text boxes.
}
\description{
This function creates a ggplot with multiple formatted text boxes.
}
\examples{
text_boxes(labels = c("[blue]First box[]", "[red]Second box[]"))

}
