% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mnb_ts.R
\name{theme_mnb_ts}
\alias{theme_mnb_ts}
\title{MNB Time Series Theme for ggplot2}
\usage{
theme_mnb_ts(base_size = 20, remove_xlab = TRUE, remove_ylab = TRUE)
}
\arguments{
\item{base_size}{Base font size for the theme. Default is inherited from ggplot2::theme_minimal().}
}
\value{
A ggplot2 theme object.
}
\description{
This function creates a custom theme for time series plots using the MNB (Magyar Nemzeti Bank) style guidelines.
}
\examples{
ggplot2::ggplot(economics, aes(x = date, y = unemploy)) +
  ggplot2::geom_line(color = "#18223e", linewidth = 1.1) +
  theme_mnb_ts()

}
