% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_memo.R
\name{use_memo}
\alias{use_memo}
\title{Create a new memo from a template}
\usage{
use_memo(filename = "memo", lang = "hu")
}
\arguments{
\item{filename}{The name of the new memo file. Default is "memo".}

\item{template}{The template to use for the new memo. Default is "mnb100".}
}
\value{
None
}
\description{
This function creates a new memo file from a specified template.
}
\examples{
use_memo("memo", "mnb100")
}
