% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_slide.R
\name{use_slide}
\alias{use_slide}
\title{Create a new slide from a template}
\usage{
use_slide(filename = "slide", template = "mnb100")
}
\arguments{
\item{filename}{The name of the new slide file. Default is "slide".}

\item{template}{The template to use for the new slide. Default is "mnb100".}
}
\value{
None
}
\description{
This function creates a new slide file from a specified template.
}
\examples{
use_slide("slide", "mnb100")
}
